move redshift tests
