+++ metadata-ingestion/tests/unit/redshift/test_redshift_config.py
